+++ tests/util/__init__.py
@@ -1,0 +1,6 @@
+"""
+tests/util package.
+
+:author: Ron Webb
+:since: 1.4.0
+"""